fix: resolve migration errors for iteration 2 features
Fixes applied:
- 0040_qr_gamification: Replace date_trunc expression index (not IMMUTABLE)
- 0040_review_helpfulness: Change calculate_wilson_score to use BIGINT
- 0040_dynamic_qr_urls: Remove invalid subquery CHECK constraint
- 0040_product_comparisons: Use o.name instead of o.display_name
- 0040_better_alternatives: Replace total_score/transparency_tier refs
- 0041_business_response: Fix notification_types INSERT columns
- 0041_review_rewards: Use app_profiles instead of profiles

All 17 migrations now apply successfully to production.

+++ supabase/migrations/0040_review_helpfulness_voting.sql
@@ -82,8 +82,8 @@ CREATE INDEX idx_verified_purchases_product ON verified_purchases(product_id);
 -- Where p = positive / total, n = total votes
 
 CREATE OR REPLACE FUNCTION calculate_wilson_score(
-    positive INTEGER,
-    negative INTEGER
+    positive BIGINT,
+    negative BIGINT
 ) RETURNS NUMERIC(10,8) AS $$
 DECLARE
     n INTEGER;
